updates to support building pre-release RPMs

+++ scripts/build-redhat.sh
@@ -28,7 +28,25 @@ cd $PKG_DIR
 cp dist/*.spec build/rpm-build/SPECS || die "Can't copy spec file."
 cp $_DIST build/rpm-build/SOURCES || die "Can't copy source package."
 
-rpmbuild --define "_topdir $PKG_DIR/build/rpm-build" -ba build/rpm-build/SPECS/*.spec || die "rpmbuild failed."
+CL_APP="$(head -n 1 dist/changelog | sed -e 's/ (.*//')"
+CL_FULL_VERSION="$(head -n 1 dist/changelog | sed -e 's/.* (//' -e 's/).*//')"
+CL_VERSION="${CL_FULL_VERSION%%-*}"
+CL_RELEASE="${CL_FULL_VERSION#*-}"
+CL_RELEASE="${CL_RELEASE//-/_}"
+CL_PKG_NAME="${CL_APP}-${CL_FULL_VERSION%-*}"
+
+EXPECTED_TARBALL="$CL_PKG_NAME.tar.gz"
+
+[[ -f "build/rpm-build/SOURCES/$EXPECTED_TARBALL" ]] || die "Expected tarball to be named [$EXPECTED_TARBALL]."
+
+rpmbuild \
+  --define "_topdir $PKG_DIR/build/rpm-build" \
+  --define "name $CL_APP" \
+  --define "version $CL_VERSION" \
+  --define "release $CL_RELEASE" \
+  --define "pkg_name $CL_PKG_NAME" \
+  -ba build/rpm-build/SPECS/*.spec \
+  || die "rpmbuild failed."
 
 mv build/rpm-build/SRPMS/* output/. || die "Can't move source RPMs."
 mv build/rpm-build/RPMS/*/* output/. || die "Can't move RPMs."
